moved smallImage into src
