+++ country-stats-app/src/SmallImage.jsx
